https://jira.catena-x.net/browse/A1ODT-127 move pages out of components
