+++ devops-tooling/dashboard/src/pages/Dashboard/Dashboard.scss
